When you discover a new unknown ID register it as used
This will permits us to recuperate at least from some failure. There is also a misuse of volatile that should have been transient. Volatile is supposed to be used in multi-threading context and transient is to be used in serialization context.

The expected effect I've had from the volatile keyword is that the field marked with it is to not be saved.