Only stock MarkID
When we were using directly the Commit object. We were creating a
big memmory retension problem since we had a big chain of commit
always depending on the previous one. With this strategy the garbage
collector will be able to remove the old Commit that have been written
and give us a some memmory during big process execution.

Figured this problem while debugging a fast-import generation.