Make sure that the files inside the recyclebin are all deleted files.
Added a unit test to prove it accurately with the first project encountered
and with it's first view. Test is inspired by the one shown in the
documentation of the recylcleBin.

It is placed in the Folder unit test because it is were the distinction is
made in the code.