Try to get rid of the janitor and moved the day by day logic inside the importer
This is a multi-step refactor. I need to simplify the code a bit because
i am now lost. The short term goal is to get back the helper on only one
to be more central and that the fast-import stream can stay the same as
long as possible. The documentation state that fast-import can live up
very long, so let take advantage of this as most as possible.