Fix time based problems when ending a importation.
Invalid state should be properly treated. For that we need to check the
type of configuration and if it is not time based, we will get the time of
the last modified item. We use the lastest because we assume that we are
now at the tip of the starteam view.