Add todo about filtering tags
This is to make sure to add filtering to tag while importing by Label. This is necessary if you want to import from starteam while the team use the label feature to atomize the commit between files and keep a consistent history. If the user want everything, this could be easily overridden by matching every thing by using the proper regular expression.