Added patch to order the revision labels
Revision labels ordering is a tricky buisness which prevent a proper history importation.
The path taken here is to use the date injected in the description of the revision label which
was generated with a CI build system. This way, we can more easily define the history based on
the automated build system.