Create a serie of commit where we can reproduce a failure
The failure we want to reproduce is the mis-match of used folder between commit re-creation in git. This can happen when a view is used to contains multiple software sources. But if the user did branching much like the way we do with subversion, we may pick the wrong subfolder and collect the commit from another branch.