Create a test to check the sub-folder importation.
This is used to reproduce a problem with the importation of subfolder. Sometime the folder that is wanted isn't the targetted folder being pick. On this test, at this version, we can drop the src/ of the sub-folder path to reproduce the problem and pick a commit from another folder. Thus removing files and adding them back the next commit.