Properly set the first past for last files
As in most case, the resume operation will not always be done on a current view of starteam but may be done on a new view where the user did create a branching point prior to the resume operation.

By doing the trick with the lastFiles and getting what is currently in the repository we make sure that the file movement will be catched and the janitor shall in the worst case delete the files as they where never available in the new view.