Provide a better support of the Revision Labels and a details history between them.
This is a squashed merge of a refactoring done for respecting the attached files added by the user.

At the same time, numerous speed improvement was done to enable a faster importation by doing less interaction with the Starteam Server. If built from the Fake-Starteam speed is increased by a lot. It should be on par if you did build against the StarteamSDK.

* Provide a means to get the file from history by view version

* Create a fake revision label

Added revision labels to the mix of required changes.

* Create test data for testing revision labels import

* Get rid of meaning less todo

* Extract the creation of a commit entry from the recursive file population

This way we can get a detailed history of the individual files to extract and
recreated a much better history

* Properly create the file based on the view version

Potential issue where the viewVersion != contentVersion.

* Create a commit strategy to experiment alternative to file population

The current limitation of readability of the code for creating commits in
GitImporter class is at its limits. The FilePopulationStrategy aim to
extract the history creation of the view since the last configuration. This
will permits some simplification on the configuration of GitImporter class
and provide a wider range of technique to extract the history from a
Starteam view.

* Implement the getViewVersion() for potential future use

* Provide a version of a file based on a label id

* Properly return null if the item is not attach to the label

* Move deletion composition and rename finding into BasePopulationStrategy

* Moved the Rename the FilePopulationStrategy

* Retab some old modified functions

* Remove some dead-code

* Add verbosity to the base strategy

* Properly build the path and fix other bug found making the tests

* Re-enable cleanup of repository after an execution of import

* Add a test for the base file population

This is to make sure that there isn't any regression added for the creation
of the list of commits.

* Make sure the shareTo is executed before the update.

* Make sure we don't ever get cached objects

* Added a means of querying the last files the strategy has seen

* Clarify the condition and reformat for understandability

* Optimize the overuse of pathname and improve testing of isChild.

* Increase test to be closer to real potential use.

* Rename class to have a better logic to it.

* Added the base integration of the commit strategy

* Cache the toString() return value for increased performance

* Finalize rename of classes and use ContentVersion instead of ViewVersion

* Correct the misleading property names.

StarteamSDK documentation isn't consistent with itself, which lead to these
kind of errors.

* Added Revision Strategy which is only a Base strategy for now

* Correct a bad assumption regarding history order

The history shall be from the newest version to the oldest

* Inform the commit strategy about the current label

And at the same time get check if we need to make a tag associated with this
label.

* Properly initialize the strategy for Change Request.

* Correct the move operation and name the commit accordingly

* Use the commit strategy in the git importer

And fix the test regarding the new behavior.

* Correct a move operation badly detected

* correct problem where openjdk != oraclejdk

* Use char sequence instead of string directly

* Replace the join with a more primitive version.

* Prevent back-and-forth for information we already know

And at the same time remove some dead-code.

* Do not enforce the missing build.date token

Fallback to the label revision time instead. If user only want a set of
build label, he will need to apply a proper filter to not select some
operationnal label.

* Provide a better explanaison of the reason we are selecting those fields.

* Extract the method which process a single file.

This is to actually override it label in the RevisionPopulationStrategy.
Some other strategy may be better to select all the files, but the simplest
for now is to simply iterate recursively through all files and only select
those that are attached to the requested labels.

* Correct bad logic regarding deleted files.

* Added basic test for revision labels strategy

* Promote debug logs to verbose

* Implement support from specific revision label support

* Create the view with the right time.

* Move set of the earliest time from GitImporter to the Base Strategy

* Simplify more the ChangeRequest commit filter

* Correct the computation of the time of the commit correction

* Add some more test for the implementation of the Revision labels

* Implement the revision labels management and importation

* Comment log relating a label with only some files

* Implement a label exclusion to rule out some misleading revisions.

* use proper javadoc syntax

* add missing header in new files.

* Cache all change request prior to starting the import

* Put the initial version in its own variable.

Such modification will enable to retreive all the files version from the
starteam repository, but with a good chance of weird effect. It is better
to keep -1 for now as the renames or move operations won't be correctly
reflected.

* Make sure to only poke getViewVersion once

It does seem to make a round-trip to the server which slow down the
conversion process.

* Add somekind of progress bar

* Provide a retry mechanism for fetching items versions.

If somebody does a commit at the same time we do make fetching of the versions
of a files, we may end up crashing. We try to protect from this crash as
from the server.

* Add missing import

* Add more resiliance when no change request availables

* Correct comment and casing of the members

* Disconnect once we are finished with the import

* Make less interaction with the Starteam Server