Fix NullPointer Exception
Add a line of defensive code to ensure that both the path returned by
the getFolderHierarchy() and the folderPattern supplied are in the
same operating system directory seperator format.

Without this change if the pattern supplied is not in Unix directory
format the folder on StarTeam is never found. As in line 531 it never
starts with the folder pattern.